- Fix a couple of wranings in the FTP code and in parsedate.y - Be safer with LC_MESSAGES - Align with the latest Zend change in call_user_function_ex()
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@26591 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/ftp/php_ftp.c
@@ -151,7 +151,7 @@ PHP_FUNCTION(ftp_connect)
 			WRONG_PARAM_COUNT;
 		}
 		convert_to_long(arg2);
-		port = arg2->value.lval;
+		port = (short) arg2->value.lval;
 		break;
 	default:
 		WRONG_PARAM_COUNT;
